requirements.txt
Django==3.0.6
djangorestframework==3.11.0
psycopg2==2.8.5
django-shortcuts==1.6
dj-database-url==0.5.0
gunicorn==20.0.4
python-decouple==3.3
whitenoise==5.1.0
Pillow==7.1.2
psycopg2-binary==2.8.5
python-dateutil==1.5
stripe==2.48.0
django-widget-tweaks==1.4.8